Update README.md
A number of people on Stack Overflow and MSDN have been confused by the localhost example in this documentation and do not realize that they need to enter their published application URL in the properties, reply URL, and web.config in order for the application to work. I have updated the ReadMe with this information.